Extend SequenceMedia to cache all media types (SVG, shader, text, binary, video)
Adds MediaCacheEntry base class and type-specific subclasses so effects
load media through SequenceMedia instead of directly from disk. This
enables embedding non-image media in .xsq files and removes effects'
dependency on xLightsFrame::CurrentDir for file resolution.

Updates ManageMediaPanel to display and manage all media types grouped
by category. Effect adjustSettings() now converts absolute paths to
relative for portability.

+++ xLights/utils/xlImage.cpp
@@ -119,7 +119,7 @@ void xlImage::Rescale(int newWidth, int newHeight) {
 
 void xlImage::Clear() {
     if (_data && _width > 0 && _height > 0) {
-        std::memset(_data.get(), 0, _width * _height * 4);
+        std::memset(_data.get(), 0, static_cast<size_t>(_width) * _height * 4);
     }
 }
